renames Monitors/BackgroundService.ts to Monitors/BackgroundRequests.ts
